pyproject.toml
[tool.poetry]
name = "coding-exercises"
version = "0.1.0"
description = ""
authors = ["Felix Quinque"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.9.2"
numpy = "^2.1.1"
loguru = "^0.7.2"
openai = "^1.51.0"
python-dotenv = "^1.0.1"
duckduckgo-search = "^6.2.13"
trafilatura = "^1.12.2"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
